Introduce the store as place to be
The store should be the main place for several objects, which are needed
by different components of the frontend with a consitent set of data.
The store itself is based on Vuex and differs in several modules,
which are also encapsulated by their namespaces. Methods (eg.
mutations, get or actions), which are based on tow or more different
modules, are assembled at the root of the store.

One module itself is crumbled also in several parts:
1. Properties, which provides a interface with the properies and their
types of the module,
2. Mutations, which define the major accesspoint for the module,
3. Getters, which are accesspoint to get values of the module,
4. Module, which define the module itself,
5. MethodNames, which define the method names of the module,
6. exceptions, which contains several exceptions which are specific for
the module,
7. Helper and other utlity classes of the module.

To provide this exception based approach this patch also introduce
several base exception types, which are stored in src/common/exceptions.
The common directory is supposed to be the place, where classes,
interfaces, etc are deposited.

So far the store contains 2 objects -> language and entity.
Entity contains a id, type. Also it contains several labels, descriptions and aliases,
which are refer to language key ( eg. en or de ). To get a specific
label of a language or something similar, the entity-module has a get[SOMETHING]ByLanguage method.
To ensure the correct type for the module you have to use the EntityStripper or a derivative of it.

Language is very similar constructed to Entity.

Bug: T207150