travis: avoid client & server build cross-influence
Ever since b0eec87bc3256bbc2112fd1611524817baf52364 the server build
seems to have influenced the build of the client JS leading to a diff
between checked-in dist folder and what travis generated (as a safety
measure to prevent us forgetting it).
This makes sure that state is not carried over from one build to the
other.

Bug: T211764