edit pen: hide if not editable
Wrap the edit pen(s) in a proprietary element which will be stripped
from the resulting mark-up (how depends on the user's edit permission).
This is not determined at rendering time based on a flag as it would
would have to be factored into the caching (hard in mw lifecycle).
This binds us to mediawiki/wikibase as it produces proprietary HTML
output that only wikibase as a middleware can understand/treat while
ordinary (modern) browsers will gracefully ignore it but would not know
what to make of it should it ever reach them directly.

This is probably not the most beautiful way to do it in vue but could
not find a cleaner way that actually worked. Tried
* a SSR-only directive https://ssr.vuejs.org/api/#directives
* a functional component https://vue-loader.vuejs.org/guide/functional.html, https://zendev.com/2018/05/07/multi-root-vue-components.html

Bug: T212136